Create locality query service

+++ src/main/java/com/partyapp/commons/entities/location/LocationDTO.java
@@ -15,6 +15,6 @@ public class LocationDTO {
     private String postalCode;
     private Long localityId;
     private String localityName;
-    private Long countryId;
+    private Integer countryId;
     private String countryName;
 }
